+++ src/solver.cpp
@@ -148,7 +148,7 @@ void Matrix::make_sparse()
 	int nz = 0; // non-null elements
 	int idx = 0;
 	for (uint i = 0; i < A_.size(); ++i) {
-		nz += A_[i].size();
+		nz += (int)A_[i].size();
 	}
 	Ax_.resize(nz);
 	Ai_.resize(nz);
@@ -156,7 +156,7 @@ void Matrix::make_sparse()
 	Ap_[0] = 0;
 	for (uint i = 0; i < A_.size(); ++i)
 	{
-		Ap_[i + 1] = Ap_[i] + A_[i].size();
+		Ap_[i + 1] = Ap_[i] + (int)A_[i].size();
 		for (column_t::iterator it = A_[i].begin();
 				it != A_[i].end(); ++it)
 		{
