+++ utils/icosahedron.cpp
@@ -297,7 +297,7 @@ void filter_mesh(vector < Triangle > & mesh,
 		Vector & p = points[i];
 
 		if (ok(p)) {
-			nums[i] = new_points.size();
+			nums[i] = (int)new_points.size();
 			new_points.push_back(p);
 		}
 	}
@@ -341,7 +341,7 @@ void divide(vector < Triangle > & mesh, vector < Vector > & points, bool save_z)
 	vector < Triangle > new_mesh;
 	vector < Vector   > new_points;
 	
-	int old_size = points.size();
+	int old_size = (int)points.size();
 	vector < map < int, int > > already(old_size);
 
 	new_points.insert(new_points.end(), points.begin(), points.end());
@@ -357,7 +357,7 @@ void divide(vector < Triangle > & mesh, vector < Vector > & points, bool save_z)
 		if (already[it->v1].find(it->v2) != already[it->v1].end()) {
 			p1 = already[it->v1][it->v2];
 		} else {
-			p1 = new_points.size();
+			p1 = (int)new_points.size();
 			Vector v1 = (a + b) / 2; v1.normalize(save_z);
 			new_points.push_back(v1);
 			
@@ -367,7 +367,7 @@ void divide(vector < Triangle > & mesh, vector < Vector > & points, bool save_z)
 		if (already[it->v1].find(it->v3) != already[it->v1].end()) {
 			p2 = already[it->v1][it->v3];
 		} else {
-			p2 = new_points.size();
+			p2 = (int)new_points.size();
 			Vector v2 = (a + c) / 2; v2.normalize(save_z);
 			new_points.push_back(v2);
 			
@@ -377,7 +377,7 @@ void divide(vector < Triangle > & mesh, vector < Vector > & points, bool save_z)
 		if (already[it->v2].find(it->v3) != already[it->v2].end()) {
 			p3 = already[it->v2][it->v3];
 		} else {
-			p3 = new_points.size();
+			p3 = (int)new_points.size();
 			Vector v3 = (b + c) / 2; v3.normalize(save_z);
 			new_points.push_back(v3);
